backend/pyproject.toml
[project]
name = "claude-nextjs-starterkit-backend"
version = "0.1.0"
description = "FastAPI 백엔드 서버"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # FastAPI 핵심 의존성
    "fastapi[standard]>=0.115.0",

    # 데이터베이스 (비동기)
    "sqlmodel>=0.0.22",
    "asyncmy>=0.2.9",
    "pymysql>=1.1.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",

    # HashiCorp Vault
    "hvac>=2.3.0",

    # 인증/보안 (✅ 공식 문서 검토 후 변경)
    "PyJWT>=2.9.0",
    "pwdlib[argon2]>=0.2.0",
    "python-multipart>=0.0.9",

    # 유효성 검사
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "email-validator>=2.2.0",

    # 유틸리티
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
